refactor: split for independent services
