--- src/rest/endpoints.js
@@ -1,7 +1,0 @@
-var exports = module.exports = {};
-
-exports.apiRoot = '/';
-
-exports.employees = {};
-exports.employees.root = exports.apiRoot + 'employees/';
-exports.employees.byId = exports.employees.root + '{0}/';
